requirements.txt
Flask==2.0.2
Jinja2==3.0.3
joblib==1.1.0
Pillow
tensorflow
numpy 
opencv-python-headless